move QuantumComponents
in helper/network/
since del star_network since it we have only 1 network
